Rename ADDRESS -> ARGUMENT in %define-opcode. Names are hard.
This value is implicitly provided in the body of all non-implied
mode instructions. For so-called "raw" instructions, it represents
the address computed by the addressing mode. For non-raw instructions,
it represents the value at that address.

I considered the word "value" but it provides no additional context.
I considered the word "operand" but it sounds slightly strange with
e.g. JMP. The point of this unhygienic implicit variable though is
to provide the argument to an opcode without the need to hand-code
use of the addressing modes. Better name suggestions welcome.