Define BVS, move BRANCH-IF, tweak raw opcodes. :thinking:
What does it mean for an opcode to be "raw"?

I came up with the concept to describe the interactions between
an addressing mode and an instruction definition. Some instructions
use the byte stored at an address where others use the address directly.

So-called "raw" instructions are simply those that use the address
(or pointer if you like) as their argument rather than the value pointed to.

The whole point of this to begin with was to allow for more simplicity in
defining the instructions. So that I could say, "Do such and such with
your argument" and have it hold true for all opcodes an instruction defines.

Thus, it makes perfect sense for any instruction whose code is more
easily _written_ using the address to be "raw", and others that use
the byte pointed to by the address to be normal. It's worth noting
that normal instructions have no way at present to receive a word.
That's because %define-opcode always calls FETCH on the address
which returns a single byte.