Fix buglet in FLAG-SET-P and add STATUS-BIT.
There are a bunch of interfaces to interacting with the status
register but there are a lot of ways we want to interact with the status.
Sometimes we want a bool to determine whether to jump, sometimes we want
the carry bit to use in addition, sometimes we want to set a flag to a value,
sometimes we want to evaluate a function to determine the value to set.

I think we've finally covered all the cases and the interface feels about right.