Fix Overflow handling for SBC. Note: I still don't fully grok this.
The algorithm to detect overflow for addition makes perfect sense.
If the inputs are both positive and the output is negative OR
if the inputs are both negative and the output is positive, overflow occurred.

That test does NOT work for subtraction. We're still trying to check
that the result of a signed operation is in -128 to 127. Based on
truth tables found online, the only cases where subtraction overflow
happens involves inputs of different signedness. For reasons I'm
unclear on, flipping the sign bit of the second operand causes the
algorithm for addition to work. Hopefully at some point I can figure
out why.

See also: http://www.righto.com/2012/12/the-6502-overflow-flag-explained.html