Add a hack to support a generic setter for raw accumulator ops.
Every time I write a 6502 emulator this gets me. My idea about passing
in the addressing modes works _except_ for Accumulator mode in
ASL, LSR, ROL, and ROR. These are the only instructions which have
some versions that operate on memory and others that operate on registers
_and_ which need to write data back in every version.

Yes, I could have used CL's generalized references/places to deal with this.
That's what I did in cl-6502, in fact. Every addressing mode had a corresponding
SETF function. But even then I needed a hack to make sure the addressing modes
that operated on memory dereferenced the byte at the address and the Accumulator
mode just used the register directly.

Besides, I like the addressing modes just _computing an address_ as they are.
Still, I should see what projects like pinky and mesen due to deal with this.