Switch from WITH-SLOTS to WITH-ACCESSORS wherever its reasonable.
It seems that WITH-SLOTS inhibits the type inference pass in SBCL.

I would hope that if the slot name was statically specified that the
compiler would lookup any type information on the struct if it knew
the object type as well but it doesn't seem to. More generally, it
seems WITH-SLOTS is intended for dynamic slot access so use
WITH-ACCESSORS even if it is slightly more verbose for a similar effect.

Still fine to use WITH-SLOTS for printer methods though.