Switch to array of structs and include skip-pc and handler.
Next step will be revising define-instruction and the various definitions.
Or maybe creating a DEFINE-INSTRUCTION* and using both initially so we can A/B
test.

DEFINE-INSTRUCTION* should create a namespaced function for the
instruction taking both CPU and ADDR-MODE as arguments. We can hardcode the
assumptions about access pattern since that is an instruction level concern, not
varying from opcode to opcode. This should be much simpler and remove the need
for %DEFINE-OPCODE altogether.

Now if only we could figure out a nice way to run Klaus 6502 tests in a proper
NES memory map. It reads and writes to regions owned by the PPU though so we'd
need to stub the fetch and store functions to make it work. :-/