Refactor CPU tests and add TEST-OPCODES-DO-NOT-ALLOCATE.
Realistically, I expect this test to possibly be a little flaky.
I've observed that successive calls to get-bytes-consed will trigger
allocation, which makes sense seeing as the result is a bignum.

The test seems to pass pretty reliably and calls to time also suggest
that no bytes are being consed during runs of NEStest which is a good
sign.

I'm seeing failures every ~20 runs. It's worth having this in the test
suite to sanity check allocation, but I would prefer it not to be part
of our CI run.