Fix braindead mistake from last night.
Good lord, how tired was I? While this produces the right result
in terms of being able to step back, I'd still like to add a unit
test and be on firmer ground as to why this specific fix is reasonable
when working when two's complement offsets for branching.

Beware that NEStest does _NO CHECKING_ of whether or not negative
branches behave as expected from what I can tell. I can throw total
garbage in that clause and it passes the test suite fine.