Add fetching for NT, AT, and PT bytes, some scrolling helpers.
This gets us about halfway to background rendering. We still need tests and
there is some remaining work in computing palette indices and drawing to the
framebuffer but the heavy lifting in terms of fetching data for visible
scanlines should be complete. It would make sense to relocate the fetching
code to the PPU package seeing how heavily it relies on PPU internals.
It's plausible that the scrolling helpers should also go in the PPU package
but I'm more open to keeping them in the rendering pass for whatever reason.