Fix the "missing scanlines" rendering bug.
Tracing the PPU address and NT bytes between scanlines indicated that we were
switching nametables at the end of each scanline in the final call to
COARSE-SCROLL-HORIZONTAL! and because nametable mirroring wasn't implemented
this led to blank tiles being fetched.

I've added preliminary support for nametable mirroring. There could definitely
be a bug in the interaction between the nametable mirroring code and the
scrolling code though. For now, this is what you get.