Add a RENDER-NAMETABLE helper and nametable rendering test.
Use manually dumped data to supply the PPU with a background to render.
RENDER-PIXEL now expects an index into the scanline for the pixel to draw.
RENDER-TILE now expects a `writer-callback` argument that will take the
computed palette index for a given pixel and cause an appropriate update.

As a result, RENDER-TILE and RENDER-PIXEL no longer are tied to a RENDERER
object and can be run against the PPU directly for debugging.

RENDER-NAMETABLE works for the limited case tested (Donkey Kong's title)
but more extensive testing should be done with a game that utilizes multiple
nametables for scrolling purposes. Fixes are probably needed.