Fix the documentation build. *FRAMEBUFFER* was too big.
It seems that if you try to export, say, a zero-initialized array of a large
enough size that PAX (or possibly 3BMD) will drop into SBCL's low-level debugger
somewhere during rendering. I'm not particularly interested in trying to figure
out what is going wrong here so for the time being I've added *FRAMEBUFFER* as a
manual export from `clones.renderer`.