Consolidate sprite-bits and bg-bits into scanline-buffer.
Having a single buffer for the scanline means that we need to account for pixel
priority while computing sprite palette indexes. The upshot is that we never
have to worry about the clearing the sprite-bits which would've been needed
otherwise and pixel priority is accounted for as soon as possible rather than at
a later step.