Fix sprite evaluation.
Turns out the bug was in EVALUATE-SPRITES. The crucial mistake was grabbing
SPRITE-BYTES using incorrect indexes which was an easy mistake to make due to
poor naming of the iteration variable.

SPRITE-INDEX wasn't the index in OAM memory of the sprite. It was the _number_
of the sprite, but because sprites are 4 byte entities you should multiply by
four to find the starting offset of the sprite.

I threw in an early return once we hit sprite overflow since there's no reason
to keep iterating through the sprites and reshuffled the creation logic a bit
but those were just minor cleanups.