Add basic MMC1 support.
There are likely still some bugs to shake out here.
This is a pretty direct port from rawbones. It's enough to get
Mega Man 2 to boot but there are definitely still issues with
both scrolling and, based on the title screen, sprite rendering.

I also set a manual width and height on the app window at 3x scale. One
other fix we needed was to supply 8k of CHR-RAM when no CHR-ROM is present.
My support for :UPPER and :LOWER in the NT-MIRROR code are also based on nothing
more than sheer intuition. It would be good to revisit those at a later date.