Fix the build, minor debug cleanups.
We were using the debug rendering helpers for the nametable tests and they
operate differently than the regular PPU rendering. This is pretty intentional
on my part. I want a method of rendering that is less concerned with how actual
frames should be drawn on the NES and more focused on the memory layout of
graphics data. But we weren't tracking scroll updates correctly in the
background iteration so it still was looking in the wrong parts of memory.

Add missing call to SYNC-HORIZONTAL-SCROLL to fix background iteration.