Remove a low-value flaky spec.
It is tempting to just use SKIP-TRIAL for now but ultimately not worth it.
This test passes locally on multiple machines for me but fails reliably in CI.

It is probably worth removing because it:
 * ties us to SBCL internals
 * cannot guarantee no other runtime behavior causes a GC
 * the negative impact of an opcode allocating should be seen quickly

I.e. If a change is made that causes the JMP instruction to allocate,
running the app or using a profiler to find the impact is probably
more effective than a failing but potentially flaky test.